requirements.txt
boto3
pandas
awswrangler
sqlalchemy-aurora-data-api
